Tests/LibWeb/Text/expected/HTML/canvas-transparent-copy-composite.txt

first draw: 255,0,0,51
transparent copy: 0,0,0,0
second draw: 255,0,0,51
